issue-61 - refactor and create new press-cards block wip
